refactor: Split functionality into separate files to allow direct imports
This allows direct imports of @sv-oss/kms-asymmetrical-jwt/lib/sign which then wont bring in the jsonwebtoken dependency. This is especially helpful for bundling
